add check by role

+++ backend/src/services/User.js
@@ -3,6 +3,11 @@ const User = require('../models/User');
 
 module.exports = {
   list: async (req, res, next) => {
+    if (req.decoded.role !== 'admin') {
+      res.status(403);
+      res.json({ status: 403, message: 'You are not allowed to make this request' });
+      return;
+    }
     try {
       const users = await User.find();
       if (!users.length) {
@@ -18,8 +23,19 @@ module.exports = {
   },
   listById: async (req, res, next) => {
     const { id } = req.params;
+    if (req.decoded.id !== id && req.decoded.role !== 'admin') {
+      res.status(403);
+      res.json({ status: 403, message: 'You are not allowed to make this request' });
+      return;
+    }
     try {
       const user = await User.findById(id);
+      if (!user) {
+        res.status(404);
+        res.json({ status: 404, message: 'No user found' });
+        return;
+      }
+      res.status(200);
       res.json(user);
     } catch (error) {
       res.json(error);
@@ -28,37 +44,38 @@ module.exports = {
   },
   create: async (req, res, next) => {
     const { name, email, password } = req.body;
-    try {
-      await User.create({ name, email, password: sha1(password) });
-      res.status(201);
-      res.json({
-        status: 201,
-        message: 'User created',
-      });
-    } catch (error) {
-      if (error.name === 'ValidationError') {
-        const messages = Object.keys(error.errors).map(key => (
-          { field: key, text: error.errors[key].message }
-        ));
-        res.status(400);
-        res.json({
-          status: 400,
-          errors: messages,
-        });
+    if (name && email && password) {
+      try {
+        await User.create({ name, email, password: sha1(password) });
+        res.status(201);
+        res.json({ status: 201, message: 'User created' });
+      } catch (error) {
+        if (error.name === 'ValidationError') {
+          const messages = Object.keys(error.errors).map(key => (
+            { field: key, text: error.errors[key].message }
+          ));
+          res.status(400);
+          res.json({ status: 400, errors: messages });
+          return;
+        }
+        res.json(error)
         return;
       }
-      res.json(error)
     }
+    res.status(400);
+    res.json({ status: 400, message: "Name or password or email can't be blank" });
     next();
   },
   delete: async (req, res, next) => {
     const { id } = req.params;
+    if (req.decoded.id !== id && req.decoded.role !== 'admin') {
+      res.status(403);
+      res.json({ status: 403, message: 'You are not allowed to make this request' });
+      return;
+    }
     try {
       await User.findByIdAndDelete(id);
-      res.json({
-        status: 200,
-        message: 'User removed',
-      });
+      res.json({ status: 200, message: 'User removed' });
     } catch (error) {
       res.json(error);
     }
